deletes some dead code, bumps the api version
